+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 -----
 - Added category by name mapping
 - Added product variation multi language support
+- Added customer main language support
+- Added manufacturer language support
 - Fixed multi language bug
 
 1.1.0